app/requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
numpy
colormath
opencv-python
